fix(ET-1792): Only start a transaction when everything is downloaded

+++ core/core-api/src/services/proton/requests.rs
@@ -53,7 +53,7 @@ pub struct GetContactsEmailsOptions {
     pub label_id: Option<LabelId>,
 
     /// Page index, i.e. the page in the resultset.
-    pub page: usize,
+    pub page: u64,
 
     /// Number of records per page.
     #[default(MAX_PAGE_ELEMENT_COUNT)]
@@ -70,7 +70,7 @@ pub struct GetContactsOptions {
     pub label_id: Option<LabelId>,
 
     /// Page index, i.e. the page in the resultset.
-    pub page: usize,
+    pub page: u64,
 
     /// Number of records per page.
     #[default(MAX_PAGE_ELEMENT_COUNT)]
